norminetted validate_map.c, refactored a function with more than 25 lines, created two extra files
