+++ src/parsing/validate_player_position.c
@@ -1,0 +1,58 @@
+#include "cub3d.h"
+
+int	ft_validate_position(t_game_data *game_data, char **map_table)
+{
+	int	i;
+	int	j;
+
+	i = (int)game_data->player.y_pos;
+	j = (int)game_data->player.x_pos;
+	if (ft_strlen(map_table[i - 1]) < (size_t)j
+		|| ft_strlen(map_table[i + 1]) < (size_t)j
+		|| ft_empty_char(map_table[i][j - 1]) == STATUS_OK
+		|| ft_empty_char(map_table[i][j + 1]) == STATUS_OK
+		|| ft_empty_char(map_table[i - 1][j]) == STATUS_OK
+		|| ft_empty_char(map_table[i + 1][j]) == STATUS_OK)
+		return (STATUS_FAIL);
+	return (STATUS_OK);
+}
+
+int	ft_validate_player_position(t_game_data *game_data, char **map_table)
+{
+	int	i;
+	int	j;
+
+	if (game_data->player.movement.direction == '0')
+		return (ft_error_msg(game_data->map_data.path, ERR_PLAYER_DIR, STATUS_FAIL));
+	i = 0;
+	while (map_table[i])
+	{
+		j = 0;
+		while (map_table[i][j])
+		{
+			if (ft_strchr("NSEW", map_table[i][j]))
+			{
+				game_data->player.x_pos = (double)j + 0.5;
+				game_data->player.y_pos = (double)i + 0.5;
+				map_table[i][j] = '0';
+			}
+			j++;
+		}
+		i++;
+	}
+	if (ft_validate_position(game_data, map_table) == STATUS_FAIL)
+		return (ft_error_msg(game_data->map_data.path, ERR_PLAYER_POS, STATUS_FAIL));
+	return (STATUS_OK);
+}
+
+int	ft_handle_player_direction(t_game_data *game_data, char current_char)
+{
+	if (ft_strchr("NSEW", current_char))
+	{
+		if (game_data->player.movement.direction != '0')
+			return (ft_error_msg(game_data->map_data.path, 
+					ERR_NUM_PLAYER, STATUS_FAIL));
+		game_data->player.movement.direction = current_char;
+	}
+	return (STATUS_OK);
+}
